requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
mongoengine==0.27.0
pymongo[srv]==3.12.3
bcrypt==4.1.2
python-dotenv==1.0.0
email_validator==2.0.0.post2
certifi==2024.8.30
